refactor(go): Move GoProperty into GoTypeMember
